--- a/.travis.yml
+++ b/.travis.yml
@@ -1,6 +1,4 @@
 node_js:
-- "4"
-- "5"
 - "6"
 sudo: false
 language: node_js