## Before:

```
                      compose
          42,098 op/s » (fn * 1)
          24,203 op/s » (fn * 2)
          13,011 op/s » (fn * 4)
           6,619 op/s » (fn * 8)
           3,414 op/s » (fn * 16)
           1,749 op/s » (fn * 32)
             888 op/s » (fn * 64)
             421 op/s » (fn * 128)
             212 op/s » (fn * 256)
             110 op/s » (fn * 512)
              54 op/s » (fn * 1024)
              
```

## After:

```
                               compose
         177,731 op/s » (fn * 1)
         175,271 op/s » (fn * 2)
         177,908 op/s » (fn * 4)
         178,933 op/s » (fn * 8)
         175,503 op/s » (fn * 16)
         175,253 op/s » (fn * 32)
         176,108 op/s » (fn * 64)
         176,910 op/s » (fn * 128)
         176,379 op/s » (fn * 256)
         159,012 op/s » (fn * 512)
         176,785 op/s » (fn * 1024)
```