@PlasmaPower 2016-07-26
Do you think you can fix the commented out tests with this implementation?