@i5ting 2016-07-26
<img width="290" alt="screen shot 2016-07-26 at 11 57 04 pm" src="https://cloud.githubusercontent.com/assets/3118295/17145479/e0c58c64-538c-11e6-9948-3ec399311b5e.png">

```
➜  compose git:(next) matcha bench/bench.js -I     
(node:25232) DeprecationWarning: child_process: options.customFds option is deprecated. Use options.stdio instead.

                      compose
         406,226 op/s » (fn * 1)
         390,251 op/s » (fn * 2)
         330,111 op/s » (fn * 4)
         414,970 op/s » (fn * 8)
         383,733 op/s » (fn * 16)
         349,785 op/s » (fn * 32)
         358,067 op/s » (fn * 64)
         404,917 op/s » (fn * 128)
         359,975 op/s » (fn * 256)
         391,651 op/s » (fn * 512)
         327,437 op/s » (fn * 1024)
         389,160 op/s » (fn * 2048)
         345,907 op/s » (fn * 4096)
         377,584 op/s » (fn * 8192)


  Suites:  1
  Benches: 14
  Elapsed: 21,467.88 ms

➜  compose git:(next) node -v
v6.0.0
```