@PlasmaPower 2016-07-26
I must say I'm very impressed by these performance improvements. I'll be happy to see how this develops :)