@fundon 2016-07-26
Results from Travis-CI

## Node V4.4.7
```
           compose
                 wait » (fn * 1)         538,324 op/s » (fn * 1)
                 wait » (fn * 2)         561,881 op/s » (fn * 2)
                 wait » (fn * 4)         559,426 op/s » (fn * 4)
                 wait » (fn * 8)         569,109 op/s » (fn * 8)
                 wait » (fn * 16)         536,691 op/s » (fn * 16)
                 wait » (fn * 32)         558,219 op/s » (fn * 32)
                 wait » (fn * 64)         549,123 op/s » (fn * 64)
                 wait » (fn * 128)         512,890 op/s » (fn * 128)
                 wait » (fn * 256)         517,420 op/s » (fn * 256)
                 wait » (fn * 512)         373,607 op/s » (fn * 512)
                 wait » (fn * 1024)         325,775 op/s » (fn * 1024)
                 wait » (fn * 2048)         446,575 op/s » (fn * 2048)
                 wait » (fn * 4096)         368,257 op/s » (fn * 4096)
                 wait » (fn * 8192)         165,344 op/s » (fn * 8192)
```