@jonathanong 2016-07-27
👍 how are the numbers with node v6?