@PlasmaPower 2016-07-27
Wait, I think `done` might still need to be fixed. I think it'll currently work with `nextFunc`, but not without.