@fundon 2016-07-27
I also create the standalone iterator https://github.com/fundon/compose-iterator. 

@PlasmaPower  Thanks a lot!