@jonathanong 2016-07-27
i want to make sure that we're not mutating the original `middleware` object but other than that LGTM. will merge tonight (ping me tomorrow if i don't haha)