@jonathanong 2016-07-28
@fundon see https://github.com/koajs/compose/pull/59#issuecomment-235780795 can you rebase and see if you can improve performance again?