@fundon 2016-07-28
@jonathanong Okay, I saw https://github.com/koajs/compose/pull/59/commits/04318f91ba3716a49e4d14cf85d3191c0ed4bcc4.