@gyson 2016-07-28
I think bench result is incorrect because `iter` object is shared by all composed function, which means `i` is also shared. When do benchmark, the shared `i` will keep increasing and make most `self[i++]` as `undefined`. Then what iterator version benched is basically noops (I think that’s why performance of thousands of middleware is just as good as single one).

It looks that the corrected iterator version is not faster than original one. from: https://github.com/koajs/compose/pull/59#issuecomment-235780795