@fundon 2016-07-28
@gyson @jonathanong Yes, I tested that the iterator version is not faster then original one.